Move analysis modules into src package
