+++ src/residual_analysis.py
