+++ tests/test_residual_integration.py
@@ -11,7 +11,7 @@ from pathlib import Path
 parent_dir = Path(__file__).parent.parent
 sys.path.insert(0, str(parent_dir))
 
-from residual_analysis import ResidualAnalyzer, ResidualMetrics
+from src.residual_analysis import ResidualAnalyzer, ResidualMetrics
 from transformer_model import TransformerModel
